8-Queens Using Goto and 2D array
Given a 8 X 8 chessboard, place 8 queens on the board, such that no two queens attack each other (in the same row, up diagonal and down diagonal). Return all distinct possible solutions for the 8 queens problem. This program uses goto statements and a 2D array data structure to represent the chessboard. Each '1' represents a Queen and '0' represents an empty slot on the chessboard.